Move common resource upload and iamge functionality to composables
